+++ client/src/App.scss
@@ -10,11 +10,14 @@ html {
             height: 100%;
             padding: 0 2rem;
             margin: 0 2rem;
-            display: flex;
-            align-items: center;
-            justify-content: center;
-            .nes-field{
-                margin: 1rem;
+            .row {
+                display: flex;
+                align-items: center;
+                justify-content: center;
+                margin: 1rem 0;
+                .nes-field{
+                    margin: 1rem;
+                }
             }
         }
     }
